start of a zillion things, not split up properly
